Combine prior and data for transition matrix

fill_transition returns the expected value of the transition matrix combining observed transitions for one time step and a prior

fill_transitions(TF, N, P = NULL, priorweight = -1, returnType = "T")

Arguments

TF
A list of two matrices, T and F, as ouput by projection.matrix.

N
A vector of observed transitions.

P
A matrix of the priors for each column. Defaults to uniform.

priorweight
total weight for each column of prior as a percentage of sample size or 1 if negative

returnType
A character vector describing the desired return value. Defaults to "T" the transition matrix.

Value

The return value depends on parameter returnType.

A - the summed matrix "filled in" using a dirichlet prior

T - just the filled in transition matrix

TN - the augmented matrix of fates -- use in calculating the CI or for simulation